extract timer into a separate class
